Fix menucolors, hilites, and revert horrible menu-glyphs kludge
After I merged Amy's "make it build on Windows" changes (which she
never intended to be merged back, as they were a "just make it work"
change, but I put her on the spot anyhow - Thanks Amy!)  A couple of
things broke (specifically hilite stairs/piles and menucolors). This
change reverts some parts of that and tidies up others.  This will
cause more problems next time a Windows build is attempted, but
hopefully we can work through these and come up with a portable
solution then.

Also got rid of this really terrible kludge I put in a while back
to attempt to communicate to window-ports that the opaque data-type
in the menu item is an object, so that mapglyph() can use it, because
menu-glyphs can't just rely on the object type, to render the glyph.
They need the material as well. Anyway, I just removed all this crap,
because there's nothing in the code that puts something on a menu with
a glyph that isn't an object, and doesn't set the identifier.a_obj.
So the opaque data type is not so opaque anymore but the interface to
the windowport remains unchanged.  If the identifier for a menu item
with a glyph is not an object, segfaults will ensue, so be careful.